+++ realm/realm-library/src/main/java/io/realm/Realm.java
@@ -1590,7 +1590,7 @@ public class Realm extends BaseRealm {
      * Deletes all objects of the specified class from the Realm.
      *
      * @param clazz the class which objects should be removed.
-     * @throws IllegalStateException if the corresponding Realm is a partially synchronized Realm, is
+     * @throws IllegalStateException if the corresponding Realm is a query-based synchronized Realm, is
      * closed or called from an incorrect thread.
      */
     public void delete(Class<? extends RealmModel> clazz) {
@@ -1728,7 +1728,7 @@ public class Realm extends BaseRealm {
      * @return a {@link RealmAsyncTask} representing a cancellable task.
      * @throws IllegalArgumentException if no {@code subscriptionName} or {@code callback} was provided.
      * @throws IllegalStateException if called on a non-looper thread.
-     * @throws UnsupportedOperationException if the Realm is not a partially synchronized Realm.
+     * @throws UnsupportedOperationException if the Realm is not a query-based synchronized Realm.
      */
     @Beta
     public RealmAsyncTask unsubscribeAsync(String subscriptionName, Realm.UnsubscribeCallback callback) {
@@ -1741,7 +1741,7 @@ public class Realm extends BaseRealm {
         }
         sharedRealm.capabilities.checkCanDeliverNotification("This method is only available from a Looper thread.");
         if (!ObjectServerFacade.getSyncFacadeIfPossible().isPartialRealm(configuration)) {
-            throw new UnsupportedOperationException("Realm is not a partially synchronized Realm: " + configuration.getPath());
+            throw new UnsupportedOperationException("Realm is fully synchronized Realm. This method is only available when using query-based synchronization: " + configuration.getPath());
         }
 
         return executeTransactionAsync(new Transaction() {
@@ -1944,7 +1944,7 @@ public class Realm extends BaseRealm {
     }
 
     /**
-     * Interface used when canceling partial sync subscriptions.
+     * Interface used when canceling query-based sync subscriptions.
      *
      * @see #unsubscribeAsync(String, UnsubscribeCallback)
      */