+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/ProcessCommitTests.java
@@ -74,6 +74,7 @@ public class ProcessCommitTests extends StandardIntegrationTest {
                 String realmUrl = Constants.SYNC_SERVER_URL;
 
                 final SyncConfiguration syncConfig = new SyncConfiguration.Builder(user, realmUrl)
+                        .fullSynchronization()
                         .modules(new ProcessCommitTestsModule())
                         .directory(getService().getRoot())
                         .build();
@@ -125,6 +126,7 @@ public class ProcessCommitTests extends StandardIntegrationTest {
         final SyncUser user = UserFactory.getInstance().createDefaultUser(Constants.AUTH_URL);
         String realmUrl = Constants.SYNC_SERVER_URL;
         final SyncConfiguration syncConfig = new SyncConfiguration.Builder(user,realmUrl)
+                .fullSynchronization()
                 .modules(new ProcessCommitTestsModule())
                 .directory(looperThread.getRoot())
                 .build();
@@ -158,6 +160,7 @@ public class ProcessCommitTests extends StandardIntegrationTest {
                 String realmUrl = Constants.SYNC_SERVER_URL;
 
                 final SyncConfiguration syncConfig = new SyncConfiguration.Builder(user, realmUrl)
+                        .fullSynchronization()
                         .modules(new ProcessCommitTestsModule())
                         .directory(getService().getRoot())
                         .name(UUID.randomUUID().toString() + ".realm")
@@ -205,6 +208,7 @@ public class ProcessCommitTests extends StandardIntegrationTest {
         final SyncUser user = UserFactory.getInstance().createDefaultUser(Constants.AUTH_URL);
         String realmUrl = Constants.SYNC_SERVER_URL;
         final SyncConfiguration syncConfig = new SyncConfiguration.Builder(user,realmUrl)
+                .fullSynchronization()
                 .modules(new ProcessCommitTestsModule())
                 .directory(looperThread.getRoot())
                 .build();