+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/objectserver/ProgressListenerTests.java
@@ -61,7 +61,9 @@ public class ProgressListenerTests extends StandardIntegrationTest {
     @Nonnull
     private SyncConfiguration createSyncConfig() {
         SyncUser user = UserFactory.createAdminUser(Constants.AUTH_URL);
-        return configFactory.createSyncConfigurationBuilder(user, Constants.SYNC_SERVER_URL).build();
+        return configFactory.createSyncConfigurationBuilder(user, Constants.SYNC_SERVER_URL)
+                .fullSynchronization()
+                .build();
     }
 
     private void writeSampleData(Realm realm) {
@@ -135,11 +137,14 @@ public class ProgressListenerTests extends StandardIntegrationTest {
         final CountDownLatch allChangesDownloaded = new CountDownLatch(1);
         SyncUser userWithData = UserFactory.createUniqueUser(Constants.AUTH_URL);
         SyncConfiguration userWithDataConfig = configFactory.createSyncConfigurationBuilder(userWithData, Constants.USER_REALM)
+                .fullSynchronization()
                 .build();
         URI serverUrl = createRemoteData(userWithDataConfig);
         SyncUser adminUser = UserFactory.createAdminUser(Constants.AUTH_URL);
 
-        final SyncConfiguration config = configFactory.createSyncConfigurationBuilder(adminUser, serverUrl.toString()).build();
+        final SyncConfiguration config = configFactory.createSyncConfigurationBuilder(adminUser, serverUrl.toString())
+                .fullSynchronization()
+                .build();
         Realm realm = Realm.getInstance(config);
         SyncSession session = SyncManager.getSession(config);
         session.addDownloadProgressListener(ProgressMode.CURRENT_CHANGES, new ProgressListener() {
@@ -164,6 +169,7 @@ public class ProgressListenerTests extends StandardIntegrationTest {
         final SyncUser userWithData = UserFactory.createUniqueUser(Constants.AUTH_URL);
         final SyncConfiguration userWithDataConfig = configFactory.createSyncConfigurationBuilder(userWithData, Constants.USER_REALM)
                 .name("remote")
+                .fullSynchronization()
                 .build();
 
         URI serverUrl = createRemoteData(userWithDataConfig);
@@ -182,6 +188,7 @@ public class ProgressListenerTests extends StandardIntegrationTest {
         SyncUser adminUser = UserFactory.createAdminUser(Constants.AUTH_URL);
         final SyncConfiguration adminConfig = configFactory.createSyncConfigurationBuilder(adminUser, serverUrl.toString())
                 .name("local")
+                .fullSynchronization()
                 .build();
         Realm adminRealm = Realm.getInstance(adminConfig);
         SyncSession session = SyncManager.getSession(adminConfig);